You can only delete an item if you created it or..
if you are an ADMIN OR DELETEITEM permissions

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -3,6 +3,7 @@ const jwt = require("jsonwebtoken");
 const { randomBytes } = require("crypto");
 const { promisify } = require("util");
 const { transport, makeANiceEmail } = require("../mail.js");
+const { hasPermission } = require("../utils.js");
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
@@ -42,8 +43,16 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // find the item
-    const item = await ctx.db.query.item({ where }, `{id title}`);
+    const item = await ctx.db.query.item({ where }, `{id title user { id }}`);
     // check if they own item or have permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ["ADMIN", "ITEMDELETE"].includes(permission)
+    );
+
+    if (!ownsItem || hasPermissions) {
+      throw new Error(`You don't have permission to do that!`);
+    }
     // delete it
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -160,6 +169,33 @@ const Mutations = {
     });
     // return new user
     return updatedUser;
+  },
+  async updatePermissions(parent, args, ctx, info) {
+    if (!ctx.request.userId) {
+      throw new Error(`You must be logged in!`);
+    }
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId
+        }
+      },
+      info
+    );
+    hasPermission(currentUser, ["ADMIN", "PERMISSIONUPDATE"]);
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions
+          }
+        },
+        where: {
+          id: args.userId
+        }
+      },
+      info
+    );
   }
 };
